requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
requests==2.32.3
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.3.0
